BUG: In Utilities/QtMobility, added patch specific to apple.
As specified here: http://doc.trolltech.com/4.3/qmake-platform-notes.html#mac-os-x

It resumes qmake is configured differently if build from the source (default to-spec macx-g++)
or distributed as a binary (default to -spec macx-xcode).

The patch make sure the option "-spec macx-g++" is passed to qmake.

Since the compileTest: QMF, NetworkManager and CoreWLAN were generating
warnings and these module doesn't seem to be required to compile
the serviceframework module, the patch will comment the corresponding lines.