ENH: Add macro ctkMacroGetAllTargetLibraries, ctkMacroGetAllCTKTargetLibraries, ctkMacroGetAllNonCTKTargetLibraries and ctkMacroShouldAddExternalProject.
1) ctkMacroGetAllTargetLibraries: Allow to retrieve all target libraries.

2) ctkMacroGetAllCTKTargetLibraries: Extract CTK libraries from 'all target libraries' (obtained using ctkMacroGetAllTargetLibraries)

3) ctkMacroGetAllNonCTKTargetLibraries: Extract non-CTK libraries from 'all target libraries'

4) ctkMacroShouldAddExternalProject: Allow to know if a given external project should be added or not.


Using macro 1) and 3), as stated in the main CMakeLists.txt, it's possible
to obtain the list NON_CTK_DEPENDENCIES

Then, in the SuperBuild.cmake, using the macro 4), it's possible to check if a given variable name
exists in the list NON_CTK_DEPENDENCIES. Such variable name could be DCMTK_LIBRARIES.

If, for example, the variable DCMTK_LIBRARIES exists in the list, the external project DCMTK is expected.

*Important*  -- Not also that within the file target_libraries.cmake located
in each Libs, Plugins or Apps subdirectory, only string should be listed. No Variable.

For example:

SET(target_libraries
  CTKCore
  DCMTK_LIBRARIES
  )

That way, the variable NON_CTK_DEPENDENCIES will contain only expected string and
not path to libraries that could be different from one plateform to another.