ENH: In superbuild.cmake, CTK-build step can be excluded if SUPERBUILD_EXCLUDE_CTKBUILD_TARGET is set to ON.
The variable should be set to ON in case a ctest driver script is
using subproject and want to configure/build/test each one of them
separately. In that specific case, we want to make sure the configure and
build error are not reported when the project is configure/build from the top
level.