ENH: QtMobility: Using the QtMobility git repository
The official git repository for QtMobility is now cloned. The patch
script checks out the 1.0 branch.

Additionally, the Unix patch for setting the PATH variable to
QT_BINARY_DIR has been removed. This lead to problems on systems which
use another name for qmake (e.g. qmake-qt4 on Fedora 12) and don't have
their Qt binary dir in the PATH variable (then the make step fails,
because the path was only set for the configure step). Now sed is used
in the patch script to replace qmake with the absolue path
(QT_QMAKE_EXECUTABLE) in the QtMobility configure script. This should
work also for custom Qt builds.