ENH: PluginFramework: Integrated the Qt Mobility service framework
The service framework from Qt Mobility can now be used to register
services and retrieve them in the same way as other PluginFramework
services (using PluginContext::getService).

The service xml file must be named servicedescriptor.xml and added
as a cached resource to the CMake ctkMacroGeneratePlugin macro. The
name of the service must be of the following form:

<plugin-symbolicname>_<version>

For example: org.commontk.myplugin_0.2.1

Additionally, the service xml file has to contain a custom property
with the key "objectclass" and the value being the class name of
the interface.