ENH: Rename ctkTitleComboBox into ctkComboBox
ctkTitleComboBox was a too limited. We replace it with ctkComboBox, a
QComboBox with a few more extra features. When no current item is selected
in the combo box, the property defaultText/defaultIcon are shown (instead
of a blank content with QComboBox). This defaultText/defaultIcon can be
forced to stay shown even if an item is selected.