COMP: Superbuild - Reviewed and simplify ExternalProject
Removed the function _ep_write_gitupdate_script which was mainly used
to re-apply patch when the repository was updated.

A better pratice consist to fixed the version/tag when patches are applied.

If nevertheless, the head of the repository is tracked and patches are
applied, the external_project_step(s) should be added.

These extra step(s) would include
 - git reset --hard HEAD
 - Switch back to master: git checkout master
 - Pull in changes: git pull
 - If a tag is specified, attempt to checkout: git checkout TAG
 - If patch_cmd is specified, attempt to re-apply the patches

.. and should depend from the "update" one