ENH: Added ctkMacroWrapPythonQt
This macro provides two ways of wrapping CTK libraries:
 1) A light wrapping
 2) A Full wrapping

Using PyhtonQt, all slots and methods added to the meta-object system
using Q_PROPERTY will be automatically available in Python. Nevertheless,
such mechanism prevent us from instantiating object or even calling
method which are not slot.



The Light wrapping takes care of wrapping both the constructor and
and destructor of class embedding the Q_OBJECT macro. Doing so allows
to instantiate and delete object from python. To enable the light wrapping,
an option entitled CTK_WRAP_PYTHONQT_LIGHT will be available after
the option CTK_LIB_Scripting/Python/Core is enabled.
Note also that the light wrapping is very fast (compare to the full wrapping).

Two log files (a)ctkMacroWrapPythonQt_log.txt and (b)ctkScriptWrapPythonQt_Light_log.txt
will be created in the current binary directory of a given library.
(a) will contain the list of file and the reason why a given class hasn't been wrapped.
(b) will contain the list of class and the constructor signature that will be wrapped.



On the other hand, the full wrapping provides a wrapping of all classes
and methods. Classes don't need to embed the Q_OBJECT macro. It relies
on the PythonQtGenerator tools which is already used by PythonQt to generate
the wrapping of the QtModules (QtGui, QtUitools, QtXml, ..). PythonQtGenerated
has been CMake'ified and will be automatically added as an external project if
CTK_WRAP_PYTHONQT_FULL is enabled. Note also that the typesystem
files used by PythonQtGenerated, which are automatically generated (within
the script ctkScriptWrapPythonQt_Full), specify just the list of classes
 to wrap (See snipped below). It means object type unknown from the meta-object system
(either returned or accepted as a method parameter) won't be handled
properly.
See http://www.pyside.org/docs/apiextractor/typesystem.html
See http://doc.trolltech.com/qtjambi-4.5.2_01/com/trolltech/qt/qtjambi-typesystem.html

Example of generated typesystem file:
 <typesystem package="org.commontk.LibName">
   <object-type name="ctkClass1"/>
   <object-type name="ctkClass2"/>
 </typesystem>



Note also that the macro will automatically exclude a file if:
 - It isn't a regular header (having .h extension)
 - It's a Pimpl header (having _p.h extenstion)
 - It contains virtual pure method

... and if CTK_WRAP_PYTHONQT_LIGHT is enabled, if:
 - It doesn't contain the Q_OBJECT macro
 - The constructor signature doesn't not match one of the following form:
     - myclass(QObject* parent)
     - myclass(QObject* parent = 0)
     - myclass(QObject* parent, Type * arg = 0)
     - myclass(QObject* parent = 0, Type * arg = 0)
     - myclass(QWidget* parent)
     - myclass(QWidget* parent = 0)
     - myclass(QWidget* parent, Type * arg = 0)
     - myclass(QWidget* parent = 0, Type * arg = 0)
     - myclass()


The option CTK_WRAP_PYTHONQT_{LIGHT or FULL} will affect all libraries.
As of today, there is no way of selectively enabled Light or Full wrapping at
either a library or file level.

The ctkMacroWrapPythonQt macro is invoked from ctkMacroBuildLib and the
variable CTK_WRAPPED_LIBRARIES_PYTHONQT is iteratively updated and will
contain the list of all CTK libraries that are wrapped with PythonQt.

If an application wants to be able to load the wrappers associated with
CTK libraries, it could linked against one or all static libraries listed by
CTK_WRAPPED_LIBRARIES_PYTHONQT. Note that the suffix PythonQt should be
added to the listed library names.

Then, the method allowing to initialize the wrapper could be called, for example,
from the overloaded method preInitialization of a custom PythonManager class
derived from ctkAbstractPythonManager.

The signature of the wrapper initialization method has the following form:
  PythonQt_init_org_commontk_<CTK_LIB_NAME>(PyObject* module)

The signature of the macro is the following:
  ctkMacroWrapPythonQt(WRAPPING_NAMESPACE TARGET SRCS_LIST_NAME SOURCES IS_WRAP_FULL)

and its different parameters are:
   WRAPPING_NAMESPACE: Namespace that should contain the library. For example: org.commontk
   TARGET ...........: Name of the wrapped library. For example: CTKWidget
   SRCS_LIST_NAME ...: Name of the variable that should contain the generated wrapper source.
                       For example: KIT_PYTHONQT_SRCS
   SOURCES ..........: List of source files that should be wrapped.
   IS_WRAP_FULL .....: Indicate if a Full wrapping if desired.