COMP: ctkMacroCompilePythonScript - Remove extra target  - Added COMMENT argument to CUSTOM_COMMAND
Having this extra target was causing problem in parallel build.

Here is the explanation of Brad king:

"The copy rules end up in Copy${kit_name}PythonFiles as dependencies
of the byte compiling custom command.
Since the two targets do not depend on one another they
can build in parallel and try to copy the same file at the same time.
Removing the unnecessary target should solve the problem."