Enable selection buttons for ctkDateRangeWidget
The widget is essentially complete for the purposes of the
the ctkDICOM application.  It could have additional options for
picking time ranges.

It would also be nice if the widget didn't resise when selecting
somthing other than "Any Date" but for now it is hiding/showing
the date selection base on context.