ctkPluginGenerator ui couldn't compile on Qt 4.6.3
The public API of ctkDirectoryButton is different depending on what Qt
version is used. This is because versions prior to 4.7.0 didn't expose the
QFileDialog::Options flags. There is no way to add preprocessing commands
to the ui file configuring the property so we need to compile a different
file depending on the version of Qt that is used.