Improve ctkMacroCompilePythonScript
To ease understanding of the parameters passed the the macro, it
now use ctkMacroParseArguments

Concept of Resources has been introduced, a python resources is a file
that will be installed along the python script without being compiled.

Using a configured script and a CustomCommand, LIBRARY_PATH is now
set before invoking python executable to compile the scripts.