Start the main test plugin in the same thread as the test functions.
The main test plugin org.commontk.pluginfwtest is now started in the same
thread as the test functions are executed. Therefore, the test suites
(QObject classes) are created in the same thread and their "initTestCase"
method is not called in parallel to the actual test methods anymore.

QTestLib only invokes "initTestCase" with Qt::AutoConnection, all other
methods are invoked with Qt::DirectConnection, who knows why...