ctkPythonConsole - Implement AutocompletePreferenceList
If a dot as been entered and if an item of possible
choices matches one of the preference list, it will be selected.

For example, in order to autocomplete the text "qt." with QPusbButton,
the list should contains: "qt.QPushButton"

By adding to the list "QPushButton" instead of "qt.QPushButton", all
module contains "QPushButton" will be autocompleted following
that preference.