Don't use the checkable header view for now
The ctkDICOMModel interface doesn't seem to have a good way
to store the StudyUID for later use in the retreival step and doesn't
have a handy way to get a checkbox status to indicate only the studies
to be downloaded.  We'll need to work on that, but for now let's just
assume that the user will be able to use all the other query tools
to be able to narrow down the specific studies to be retrieved (i.e.
by date, name, modality, etc).  In practice this seems to be
pretty workable.