ctkCallback now accepts callback data
By default, the data passed to the associated function will be a pointer
to the ctkCallback itself. Using property/setProperty it's then
possible to associate data to the callback itself.

It's also possible to set a custom data structure using setCallbackData.

Note that this commit is breaking backward compatibility. Existing code
will have to add an extra "void*" parameter to the function associated to
the callback.