Make the default storage port different than the default query port
Making these different allows the creation of a local instance
of dcmqrscp (running on the same host) that is able to support
query/retrieve with the ctkDICOM code without requiring any changes
on the part of the user.  Without this change, it would be trying to
reuse the same ports.