ctkSettings - Properly handle empty QStringList
The method Property::setValue was checking the return value of QObject::setProperty.
Nevertheless, if an empty QStringList is stored in the settings and read back,
 the associated QVariant will be invalid.

In that specific case, if the type of the associated property is QStringList
and it the value to set is Invalid, we ensure an empty QStringList is set.

A QtBug has also been reported. See http://bugreports.qt.nokia.com/browse/QTBUG-19823