Superbuild - Fix CTK-Build step
In ExternalProject.cmake, when using ExternalProject_Add, if the
command associated with the build step is not specified. One is generated
based on the configure command. In our case, since the configure
command had been explicitly disabled (set to an empty string), the build
command was then defaulting to 'make'.

Let's add a dummy configure step so that BUILD_COMMAND is properly setup.

For more details, see _ep_get_configure_command_id
and _ep_get_build_command in ExternalProject.cmake