ctkDoubleSlider::setRange wasn't bounding the value
In ctkDoubleSlider::setRange():
In some cases (where calling d->Slider->setRange(...) wasn't changing the
QSlider value), d->Value was keeping a value that was outside of the
new range.
Indeed, in the case (-1, -0.5, 1) (min, value, max) the QSlider have the
following values [-1, 0, 1]. Changing the range to [0., 1.] is changing the
QSlider range to [0, 1] which doesn't change the value (0).