ctkIconEnginePlugin::keys returns ctkIconEngine
When saving a QIcon into a QDataStream, the icon engine key is used.
Later, when loading a QIcon from a QDataStream, that same key is used to
find the icon engine used.
ctkPixmapIconEngine was the key saved but couldn't be retrieved because
the plugin wasn't exposing it.
You might have to clear your Trolltech registry (Qt Factory Cache*) in
order for the ctkIconEnginePlugin meta information to be reloaded