Added the errorTextEnabled property of a ctkWorkflowGroupBox.
If errorTextEnabled is turned on, then a possible error text
will be shown just underneath the client area.
If it is turned off, then a possible error text will never
be shown. By default, this property is ON.

To enable python access to the groupbox of a workflowWidget, the
workflowGroupBox() method was marked as Q_INVOKABLE.