BUG: ctkWorkflow / ctkWorkflowStep - Fix memory leak
Delete Workflow setter in WorkflowStep as long as we don't want the user to create unassociated steps.
It avoids having trouble when cleaning between WorflowStep & WotkflowStepWidget.
Create a list of registered worflowStep in workflow_p to manage the cleaning.
Every time a workflowStep (but not workflowStepWidget) is created, it registrates in the workflow.
The workflow manages then the destruction of the steps when he calls his dtor.