Expose suspend and resume methods for the dicom model
It seems that while the dicom model is active it keeps open
QSqlQuery instances that have a negative impact on the ability
to do inserts into the database.  These new methods allow a
user of this class to turn off the database and turn it back
on again without having to re-specify the directory.

There may be a faster way to implement this (i.e. by calling
finish on the queries and then re-execing them later) but this
implementation is much faster than what existed previously, where
the sqlite database would sleep waiting for a chance to insert
new data.