Do not use hidden visibility for gcc < 4.5.
With gcc < 4.5, RTTI symbols from classes declared in third-party
libraries which are not "gcc visibility aware" are marked with
hidden visibility in DSOs which include the class declaration and
which are compiled with hidden visibility. This leads to
dynamic_cast and exception handling problems. While this problem
could be worked around by sandwiching the include directives for
the third-party headers between "#pragma visibility push/pop"
statements, it is generally safer to just use default visibility
with gcc < 4.5.