Add Testing utilities to do unit testing with QtTest framework
Within the CTest framework, QTEST_MAIN can't be used as it defines a
duplicated main() function. CTest requires the test function to have
the same name as the test file.
QTest::mouseMove() is not supporting moves whith a button down.
We add a utility function ctkTest::mouseMove() to support the feature.
It is surprising that such behavior doesn't exist by default in Qt. There
might be some associated issues with it.