During resolving a plug-in, set the state of resolved dependencies.
When resolving a plug-in, its dependencies are resolved too. We now
immediately set their state to RESOLVED and send the appropriate event.
This seems okay - the OSGi specs state that the framework may try to
resolve plug-ins at any time. This change ensures that resolve events
are fired according to plug-in dependency order. Should fix #127.