ctkMessageBox::setDontShowAgainSettingsKey() turns visibility on
When the key is set, it's most likely that the checkbox should be visible.
This simplifies the setup of ctkMessageBox.
If the default behavior is unwanted, it's still possible to set the
 visibility back after setting the key.

The order of properties is changed, just in case ctkMessageBox is configured
using Qt Designer, it would set the properties in the same order.
Changing the DontShowAgainSettingsKey property should be done first, so that
DontShowAgainVisible can then override the default behavior.