Fix title height of ctkCollapsibleGroupBox
A Qt bugs makes it so the checkbox height is not taken into account for
the height of the title SC_GroupBoxContents (see QTBUG #6326)
This is especially a problem with the Mac style.

PE_IndicatorArrowDown doesn't have a pixel metric, the current font height
is used instead of the checkbox height.

Another fix would be to take into account the height of the checkbox by
reimplementing subElementRect for SC_GroupBoxContents and SC_GroupBoxFrame