ctkFDHandler - The polling thread now terminates graciously
Before that change the thread in charge of reading the content
of either stderr or stdout descriptor was terminated using:

  this->terminate()
  this->wait();

On some operating system, using this strategy was leaving the thread in
a inconsistent state. See http://na-mic.org/Mantis/view.php?id=1537

The approach implemented here is less brutal. Indeed, the "Enabled" boolean
variable is protected by a QMutex and allows both the main thread and
the polling thread to either set or get the Enabled variable in a
thread-safe way.