BUG: avoid deadlock on mac on file import
On mac, bringing up the progress dialog on top of the
file dialog leads to a deadlock where the
dialogs go away but the underlying widget is not
responsive and the application requires a Force Quit.
Avoid this by making the progress dialog non-modal.
Unfortunately it remains behind the file dialog but this
is better than hanging.