Fix segfaults associated with ctkSimplePythonShell tests
* In both "wrappedVTKQInvokableTest.py" and "wrappedVTKSlotTest.py",
a vtkTable was instantiated within python and set to wrapped cpp
instance of either (a) "ctkTestWrappedVTKQInvokable" or (b) "ctkTestWrappedVTKSlot".

Within each on of these classes (a and b), the "Delete" function
was called unconditionally on the MyTable private member. Two problems
were happening:

 - the first instance of vtkTable initially created within the constructor
of classes a and b was leaked.

 - the vtkTable instantiated within python referenced with the "t2"
variable was set to the classes a and b. Then, when python was finalized
it was deleting the object allocated within python. Next, in addition the
Delete function was also called within the destructor of classes a and b.
The pointer being invalid, a seg fault occurred. Using vtkSmartPointer
help us to elegantly address this issue.