ctkLayoutManager was not properly cleaning the layout
clearLayout() was not deleting the layout elements (e.g. Splitters, QTabWidget)
created previously by setupLayout().
It was a problem when reparenting the views when applying the new layout in the viewport in setupLayout.
The reparenting of a QTabWidget child was setting the WA_WState_ExplicitShowHide flag on the widgets which
was preventing _q_showIfNotHidden() from setting the visibility of the child back at the next update.
Moreover the QTabWidgets and QSplitters were not destroyed until ctkLayoutManager was destroyed.
When clearing layout, it is enforced that the widgets don't redraw (QWidget::setUpdatesEnabled(false)) and the layouts don't resize (QLayout::setEnabled(false))to prevent flickering.