Add schema updating to ctkDICOM code
This is based on work by Marco Nolden to do an in-place update
of the database schema.

The new commits here provide a way to detect when a schema
update is needed using a new table in the database.

Also the schema update emits signals as the update goes on.

Also there is now a test and some examples of old schema so
we can verify things are working.

Also there is a uniform method to reset the effiency variables
so they don't accidentally result in missed entries after
removing data or updating the schema.

The app widget and ctkDICOM example application now support updating
the schema as neeed.