Debug and release libs are now stored separately
Changed the DCMTK script to install the debug and release libs as well as
the binaries to Debug/Release subfolders. Furthermore changed the
FindDCMTK scipt to search these folders and add only the libs
corresponding to the current build selection.

This was done to fix the inability to do a Microsoft Visual Studio 2010
Debug and Release superbuild.