The "channel" element is now optionally allowed for all parameters.
Previously, it was not possible to set the channel to "output" for
simple parameter types. Now, channel is optional for all types and
defaults to "input" (when parsing the XML description). Trying to do
a "restriction" on the minOccurs attribute of the channel element in
derived schema types turned out to be too complicated.