Ensure ctkAbstractPythonManager::executeFile set __file__
The code has also been refactored to ensure sys.path is cleaned in case of
error.

Tests have been added.

A new method named "resetErrorFlag" has been added to both CTK and PythonQt.
This method should be called before any piece of code that ends up
calling "handleError" in case of error. It ensures that the errorOccured
method won't return True when it is not expected.

See #244