Ensure ctkFDHandler can be enabled/disabled multiple times
To give the opportunity to the thread to exit the polling loop, the
character "\n" is now written when disabling the handler.

The pipe is also setup each time the thread is enabled. Additionally,
the reading side of the redirection pipe is also closed when the thread is
disabled.

Tests have also been updated to enable/disable the handler multiple times
in a row.

See Slicer issue 2535