Added screenshot output resolution support
The user can now select whether to scale the current viewport or to get
a fixed resolution screenshot of the viewport. The lock aspect ratio
button, when selected, makes sure the screenshot image has the same
aspect ratio as the viewport.

ENH: Added method to enforce resolution for output screenshot