Relayout ctkThumbnailListWidget when thumbnails become visible
ctkThumbnailListWidget controls it's own scrollarea internal widget size.
The size is manually computed from the layout (ctkFlowLayout).
However, it fails at computing a correct size if the added widgets are
not visible at the time of the relayout.
ctkThumbnailListWidget now observes the showevent of such widgets to
trigger a relayout.
Closes #272