Add option allowing to reduce number of python copy/compilation targets
By globally defining the variable CTK_COMPILE_PYTHON_SCRIPTS_GLOBAL_TARGET_NAME to a
non-empty string or by specifying the macro option 'GLOBAL_TARGET',
the following targets will be defined for the whole build system:
 - Copy<GLOBAL_TARGET_NAME>PythonResourceFiles
 - Copy<GLOBAL_TARGET_NAME>PythonScriptFiles
 - Compile<GLOBAL_TARGET_NAME>PythonFiles

For complex projects, this can help reducing the number of targets and
simplify the manual rebuild of copy and compile targets.

Also make the custom command in charge of compiling the script dependent
of Copy${target}PythonScriptFiles target. That building a compile target
implies the associated copy target will be executed.