ENH: Add ctkSpinBox DecimalsOption::UseShortcut
ctkSpinBox::DecimalsOption controls how decimals are handled. In the
regular (Fixed) mode, the spinbox behaves like a QDoubleSpinBox, i.e.
the number of decimals is set by setDecimals().
With the UseShortcut flag, the ctkSpinBox responds to keyboard shortcuts to
add (CTRL +) decimals, remove (CTL -) decimals and restore the
default number of decimals (CTRL 0) initially set by setDecimals().

This led to the update of the ctk widgets that use the ctkSpinBox. For most
of them it's pretty straightforward as they can listen to the
decimalsChanged(int) event and update themselves.
In the case of the slider widget, the AutoSpinBoxWidth was extended to
support the change of width due to a change in decimals as well as the
change in decimals (that is propagated to the siblings).

See Slicer issue #2480