Fix ctkSliderWidget assert when changing decimals with DecimalsByValue
By blocking signals in ctkSliderWidget::changeValue, the change of decimals
(because of ctkDoubleSpinBox::DecimalsByValue) is not catched by
ctkSliderWidget and not propagated to the slider which causes a discrepancy
with the minimum/maximum values between the spinbox and the slider.
The test values come from a real case that was asserting in
ctkSliderWidget.