Do not append non-existent directory to "*_{INCLUDE,LIBRARY}_DIRS" variables
This is particularly useful when building against library that are not
always defining a "<project_name>_LIBRARY_DIRS" variable. This commit
avoids appending a variable name to the list of possible directories
that should be considered.

For example, within ITKv4, the variable ITK_LIBRARY_DIRS is not
always set. See here for more details: https://issues.itk.org/jira/browse/ITK-3026