If given a new or newly modified file, update the Images database row
Fixes #357 and is verified by ctkDICOMDatabaseTest5.

When the user imports a file that is already in the database, delete
the old row for that image and let it be re-imported so that the
database will always point to the most current location of the file
(or at least the most recently imported).